Add logo in SVG format
This was created by hand in Inkscape to match the Uiua-generated
PNG as closely as possible. I made it originally to submit to
vscode-icons for inclusion, but having it tracked here in the repo as
the "canonical" SVG version should be useful when adding the logo
to desktop environments or custom launchers etc.

It was carefully crafted to reduce the size to ~3.5kB.

The version for vscode-icons was further golfed and minified to be <2kB,
see https://github.com/vscode-icons/vscode-icons/pull/3358.