fix(editor): common macos shortcuts for next and previous respected
Problem:

On macOS, next and previous tab for documents/tabs within applications
are bound to Cmd-Shift-[ and Cmd-Shift-]. Pressing these while in the
editor changes tabs but invokes the delimiter insertion code.

Solution:

Only run the delimiter handling code if the `Cmd` button is not held
pressed.  This is detectable via use of the `meta` key in the browser.