Add support for hidden code in website embedded editors
When a code block for the website contains `# ^^^ HIDDEN ^^^` in the
markdown file, all code above the comment will not be shown in the
embedded editor, but will still take effect. Displayed line numbers are
offset accordingly, and clicking the "copy link" button will give a url
containing all code, including the hidden code.